wip testing continuous profiling implemenation so far
