+++ Tests/SentryProfilerTests/SentryLegacyProfilerTests.swift
@@ -4,237 +4,9 @@ import SentryTestUtils
 import XCTest
 
 #if os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
-class SentryProfilerSwiftTests: XCTestCase {
-    private static let dsnAsString = TestConstants.dsnAsString(username: "SentryProfilerSwiftTests")
+class SentryLegacyProfilerTests: XCTestCase {
 
-    private class Fixture {
-        lazy var options: Options = {
-            let options = Options()
-            options.dsn = SentryProfilerSwiftTests.dsnAsString
-            return options
-        }()
-        lazy var client: TestClient? = TestClient(options: options)
-        lazy var hub: SentryHub = {
-            let hub = SentryHub(client: client, andScope: scope)
-            hub.bindClient(client)
-            return hub
-        }()
-        let scope = Scope()
-        let message = "some message"
-        let transactionName = "Some Transaction"
-        let transactionOperation = "Some Operation"
-
-        lazy var systemWrapper = TestSentrySystemWrapper()
-        lazy var processInfoWrapper = TestSentryNSProcessInfoWrapper()
-        lazy var dispatchFactory = TestDispatchFactory()
-        var metricTimerFactory: TestDispatchSourceWrapper?
-        lazy var timeoutTimerFactory = TestSentryNSTimerFactory()
-        let dispatchQueueWrapper = TestSentryDispatchQueueWrapper()
-
-        let currentDateProvider = TestCurrentDateProvider()
-
-#if !os(macOS)
-        lazy var displayLinkWrapper = TestDisplayLinkWrapper(dateProvider: currentDateProvider)
-        lazy var framesTracker = SentryFramesTracker(displayLinkWrapper: displayLinkWrapper, dateProvider: currentDateProvider, dispatchQueueWrapper: SentryDispatchQueueWrapper(), keepDelayedFramesDuration: 0)
-#endif // !os(macOS)
-
-        init() {
-            SentryDependencyContainer.sharedInstance().dateProvider = currentDateProvider
-            options.profilesSampleRate = 1.0
-            options.tracesSampleRate = 1.0
-
-            SentryDependencyContainer.sharedInstance().systemWrapper = systemWrapper
-            SentryDependencyContainer.sharedInstance().processInfoWrapper = processInfoWrapper
-            dispatchFactory.vendedSourceHandler = { eventHandler in
-                self.metricTimerFactory = eventHandler
-            }
-            SentryDependencyContainer.sharedInstance().dispatchFactory = dispatchFactory
-            SentryDependencyContainer.sharedInstance().timerFactory = timeoutTimerFactory
-            SentryDependencyContainer.sharedInstance().dispatchQueueWrapper = dispatchQueueWrapper
-            
-            SentryDependencyContainer.sharedInstance().random = TestRandom(value: 0.5)
-
-            systemWrapper.overrides.cpuUsage = NSNumber(value: mockCPUUsage)
-            systemWrapper.overrides.memoryFootprintBytes = mockMemoryFootprint
-            systemWrapper.overrides.cpuEnergyUsage = 0
-
-#if !os(macOS)
-            SentryDependencyContainer.sharedInstance().framesTracker = framesTracker
-            framesTracker.start()
-            displayLinkWrapper.call()
-#endif // !os(macOS)
-        }
-
-        /// Advance the mock date provider, start a new transaction and return its handle.
-        func newTransaction(testingAppLaunchSpans: Bool = false, automaticTransaction: Bool = false, idleTimeout: TimeInterval? = nil) throws -> SentryTracer {
-            let operation = testingAppLaunchSpans ? SentrySpanOperationUILoad : transactionOperation
-
-            if automaticTransaction {
-                return hub.startTransaction(
-                    with: TransactionContext(name: transactionName, operation: operation),
-                    bindToScope: false,
-                    customSamplingContext: [:],
-                    configuration: SentryTracerConfiguration(block: {
-                        if let idleTimeout = idleTimeout {
-                            $0.idleTimeout = idleTimeout
-                        }
-                        $0.waitForChildren = true
-                        $0.timerFactory = self.timeoutTimerFactory
-                    }))
-            }
-
-            return try XCTUnwrap(hub.startTransaction(name: transactionName, operation: operation) as? SentryTracer)
-        }
-
-        // mocking
-
-        let mockCPUUsage = 66.6
-        let mockMemoryFootprint: SentryRAMBytes = 123_455
-        let mockEnergyUsage: NSNumber = 5
-        let mockUsageReadingsPerBatch = 3
-
-#if !os(macOS)
-        // SentryFramesTracker starts assuming a frame rate of 60 Hz and will only log an update if it changes, so the first value here needs to be different for it to register.
-        let mockFrameRateChangesPerBatch: [FrameRate] = [.high, .low, .high, .low]
-
-        // Absolute timestamps must be adjusted per span when asserting
-        var expectedSlowFrames = [[String: Any]]()
-        var expectedFrozenFrames = [[String: Any]]()
-        var expectedFrameRateChanges = [[String: Any]]()
-
-        func resetGPUExpectations() {
-            expectedSlowFrames = [[String: Any]]()
-            expectedFrozenFrames = [[String: Any]]()
-            expectedFrameRateChanges = [[String: Any]]()
-        }
-#endif // !os(macOS)
-
-        func gatherMockedMetrics(span: Span) throws {
-            // clear out any errors that might've been set in previous calls
-            systemWrapper.overrides.cpuUsageError = nil
-            systemWrapper.overrides.memoryFootprintError = nil
-            systemWrapper.overrides.cpuEnergyUsageError = nil
-
-            // gather mocked metrics readings
-            for _ in 0..<mockUsageReadingsPerBatch {
-                self.metricTimerFactory?.fire()
-
-                // because energy readings are computed as the difference between sequential cumulative readings, we must increment the mock value by the expected result each iteration
-                systemWrapper.overrides.cpuEnergyUsage = NSNumber(value: systemWrapper.overrides.cpuEnergyUsage!.intValue + mockEnergyUsage.intValue)
-            }
-
-#if !os(macOS)
-            var shouldRecordFrameRateExpectation = true
-            
-            func changeFrameRate(_ new: FrameRate) {
-                displayLinkWrapper.changeFrameRate(new)
-                shouldRecordFrameRateExpectation = true
-            }
-            
-            func renderGPUFrame(_ type: GPUFrame) {
-                switch type {
-                case .normal:
-                    let currentSystemTime: UInt64 = currentDateProvider.systemTime()
-                    print("expect normal frame to start at \(currentSystemTime)")
-                    displayLinkWrapper.normalFrame()
-                case .slow:
-                    let duration = displayLinkWrapper.middlingSlowFrame().toNanoSeconds()
-                    let currentSystemTime = currentDateProvider.systemTime()
-                    print("will expect \(String(describing: type)) frame starting at \(currentSystemTime)")
-                    expectedSlowFrames.append([
-                        "elapsed_since_start_ns": String(currentSystemTime),
-                        "value": duration
-                    ])
-                case .frozen:
-                    let duration = displayLinkWrapper.fastestFrozenFrame().toNanoSeconds()
-                    let currentSystemTime = currentDateProvider.systemTime()
-                    print("will expect \(String(describing: type)) frame starting at \(currentSystemTime)")
-                    expectedFrozenFrames.append([
-                        "elapsed_since_start_ns": String(currentSystemTime),
-                        "value": duration
-                    ])
-                }
-                if shouldRecordFrameRateExpectation {
-                    shouldRecordFrameRateExpectation = false
-                    let currentSystemTime = currentDateProvider.systemTime()
-                    print("will expect frame rate \(displayLinkWrapper.currentFrameRate.rawValue) at \(currentSystemTime)")
-                    expectedFrameRateChanges.append([
-                        "elapsed_since_start_ns": String(currentSystemTime),
-                        "value": NSNumber(value: displayLinkWrapper.currentFrameRate.rawValue)
-                    ])
-                }
-            }
-            
-            /*
-             * Mock a series of GPU frame renders of varying quality (normal/slow/frozen) and
-             * refresh rate changes. The refresh rate changes ("|") happen at the same time as
-             * the frame render they appear above. Time is not to scale; frozen frames last
-             * much longer than the lower end of slow frames.
-             *
-             * refresh rate:  |---60hz--------------|---120hz------|---60hz--------------------------------------|
-             * time:          N--S----N--F----------N-N-S--N-F-----N--N--S----N--F----------N--S----N--F----------
-             */
-            changeFrameRate(.low)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.slow)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.frozen)
-            renderGPUFrame(.normal)
-            changeFrameRate(.high)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.slow)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.frozen)
-            renderGPUFrame(.normal)
-            changeFrameRate(.low)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.slow)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.frozen)
-            renderGPUFrame(.normal)
-            changeFrameRate(.high)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.slow)
-            renderGPUFrame(.normal)
-            renderGPUFrame(.frozen)
-#endif // !os(macOS)
-
-            // mock errors gathering cpu usage and memory footprint and fire a callback for them to ensure they don't add more information to the payload
-            systemWrapper.overrides.cpuUsageError = NSError(domain: "test-error", code: 0)
-            systemWrapper.overrides.memoryFootprintError = NSError(domain: "test-error", code: 1)
-            systemWrapper.overrides.cpuEnergyUsageError = NSError(domain: "test-error", code: 2)
-            metricTimerFactory?.fire()
-
-            // clear out errors for the profile end sample collection
-            systemWrapper.overrides.cpuUsageError = nil
-            systemWrapper.overrides.memoryFootprintError = nil
-            systemWrapper.overrides.cpuEnergyUsageError = nil
-        }
-
-        // app start simulation
-
-        lazy var appStart = currentDateProvider.date()
-        lazy var appStartSystemTime = currentDateProvider.systemTime()
-        var appStartDuration = 0.5
-        lazy var appStartEnd = appStart.addingTimeInterval(appStartDuration)
-
-#if !os(macOS)
-        func getAppStartMeasurement(type: SentryAppStartType, preWarmed: Bool = false) -> SentryAppStartMeasurement {
-            let runtimeInitDuration = 0.05
-            let runtimeInit = appStart.addingTimeInterval(runtimeInitDuration)
-            let mainDuration = 0.15
-            let main = appStart.addingTimeInterval(mainDuration)
-            let didFinishLaunching = appStart.addingTimeInterval(0.3)
-            appStart = preWarmed ? main : appStart
-            appStartDuration = preWarmed ? appStartDuration - runtimeInitDuration - mainDuration : appStartDuration
-            appStartEnd = appStart.addingTimeInterval(appStartDuration)
-            return SentryAppStartMeasurement(type: type, isPreWarmed: preWarmed, appStartTimestamp: appStart, runtimeInitSystemTimestamp: appStartSystemTime, duration: appStartDuration, runtimeInitTimestamp: runtimeInit, moduleInitializationTimestamp: main,
-                                             sdkStartTimestamp: appStart, didFinishLaunchingTimestamp: didFinishLaunching)
-        }
-#endif // !os(macOS)
-    }
-
-    private var fixture: Fixture!
+    private var fixture: SentryProfileTestFixture!
 
     override class func setUp() {
         super.setUp()
@@ -243,7 +15,7 @@ class SentryProfilerSwiftTests: XCTestCase {
 
     override func setUp() {
         super.setUp()
-        fixture = Fixture()
+        fixture = SentryProfileTestFixture()
     }
 
     override func tearDown() {
@@ -270,7 +42,7 @@ class SentryProfilerSwiftTests: XCTestCase {
         self.fixture.currentDateProvider.advanceBy(nanoseconds: 1.toNanoSeconds())
         span.traceId = SentryId()
         span.finish()
-        try self.assertValidProfileData()
+        try self.assertValidLegacyProfileData()
     }
 
     func testConcurrentProfilingTransactions() throws {
@@ -294,7 +66,7 @@ class SentryProfilerSwiftTests: XCTestCase {
                 fixture.currentDateProvider.advanceBy(nanoseconds: 100)
             }
 
-            let threadMetadata = ThreadMetadata(id: 1, priority: 2, name: "test-thread")
+            let threadMetadata = SentryProfileTestFixture.ThreadMetadata(id: 1, priority: 2, name: "test-thread")
             try addMockSamples(threadMetadata: threadMetadata)
 
             for (i, span) in spans.enumerated() {
@@ -303,7 +75,7 @@ class SentryProfilerSwiftTests: XCTestCase {
                 span.finish()
                 XCTAssertEqual(SentryLegacyProfiler.currentProfiledTracers(), UInt(numberOfTransactions - (i + 1)))
 
-                try self.assertValidProfileData(expectedThreadMetadata: [threadMetadata])
+                try self.assertValidLegacyProfileData(expectedThreadMetadata: [threadMetadata])
 
                 // this is a complicated number to come up with, see the explanation for each part...
                 let expectedUsageReadings = fixture.mockUsageReadingsPerBatch * (i + 1) // since we fire mock metrics readings for each concurrent span,
@@ -341,7 +113,7 @@ class SentryProfilerSwiftTests: XCTestCase {
         let spanA = try fixture.newTransaction()
         fixture.currentDateProvider.advanceBy(nanoseconds: 1.toNanoSeconds())
         let expectedAddressesA: [NSNumber] = [0x1, 0x2, 0x3]
-        let expectedThreadMetadataA = ThreadMetadata(id: 1, priority: 2, name: "test-thread1")
+        let expectedThreadMetadataA = SentryProfileTestFixture.ThreadMetadata(id: 1, priority: 2, name: "test-thread1")
         try addMockSamples(threadMetadata: expectedThreadMetadataA, addresses: expectedAddressesA)
 
         // time out profiler for span A
@@ -354,43 +126,43 @@ class SentryProfilerSwiftTests: XCTestCase {
         let spanB = try fixture.newTransaction()
         fixture.currentDateProvider.advanceBy(nanoseconds: 0.5.toNanoSeconds())
         let expectedAddressesB: [NSNumber] = [0x7, 0x8, 0x9]
-        let expectedThreadMetadataB = ThreadMetadata(id: 4, priority: 5, name: "test-thread2")
+        let expectedThreadMetadataB = SentryProfileTestFixture.ThreadMetadata(id: 4, priority: 5, name: "test-thread2")
         try addMockSamples(threadMetadata: expectedThreadMetadataB, addresses: expectedAddressesB)
 
         // finish span B and check profile data
         spanB.finish()
-        try self.assertValidProfileData(expectedAddresses: expectedAddressesB, expectedThreadMetadata: [expectedThreadMetadataB])
+        try self.assertValidLegacyProfileData(expectedAddresses: expectedAddressesB, expectedThreadMetadata: [expectedThreadMetadataB])
 
         // finish span A and check profile data
         spanA.finish()
-        try self.assertValidProfileData(expectedAddresses: expectedAddressesA, expectedThreadMetadata: [expectedThreadMetadataA])
+        try self.assertValidLegacyProfileData(expectedAddresses: expectedAddressesA, expectedThreadMetadata: [expectedThreadMetadataA])
     }
 
     func testProfileTimeoutTimer() throws {
-        try performTest(shouldTimeOut: true)
+        try performLegacyProfilingTest(shouldTimeOut: true)
     }
 
     func testStartTransaction_ProfilingDataIsValid() throws {
-        try performTest()
+        try performLegacyProfilingTest()
     }
 
     func testProfilingDataContainsEnvironmentSetFromOptions() throws {
         let expectedEnvironment = "test-environment"
         fixture.options.environment = expectedEnvironment
-        try performTest(transactionEnvironment: expectedEnvironment)
+        try performLegacyProfilingTest(transactionEnvironment: expectedEnvironment)
     }
 
 #if !os(macOS)
     func testProfileWithTransactionContainingStartupSpansForColdStart() throws {
-        try performTest(uikitParameters: UIKitParameters(launchType: .cold, prewarmed: false))
+        try performLegacyProfilingTest(uikitParameters: UIKitParameters(launchType: .cold, prewarmed: false))
     }
 
     func testProfileWithTransactionContainingStartupSpansForWarmStart() throws {
-        try performTest(uikitParameters: UIKitParameters(launchType: .warm, prewarmed: false))
+        try performLegacyProfilingTest(uikitParameters: UIKitParameters(launchType: .warm, prewarmed: false))
     }
 
     func testProfileWithTransactionContainingStartupSpansForPrewarmedStart() throws {
-        try performTest(uikitParameters: UIKitParameters(launchType: .warm, prewarmed: true))
+        try performLegacyProfilingTest(uikitParameters: UIKitParameters(launchType: .warm, prewarmed: true))
     }
 #endif // !os(macOS)
 
@@ -399,7 +171,7 @@ class SentryProfilerSwiftTests: XCTestCase {
         fixture.hub.configureScope { scope in
             scope.setEnvironment(expectedEnvironment)
         }
-        try performTest(transactionEnvironment: expectedEnvironment)
+        try performLegacyProfilingTest(transactionEnvironment: expectedEnvironment)
     }
 
     func testStartTransaction_NotSamplingProfileUsingEnableProfiling() throws {
@@ -552,27 +324,9 @@ class SentryProfilerSwiftTests: XCTestCase {
         XCTAssertEqual(self.fixture.client?.captureEventWithScopeInvocations.count, 0)
     }
 #endif // !os(macOS)
-    
-    // given a profiler initialized with legacy mode
-        // ensure a timeout timer is set
-        // ensure a background notification observer is set
-        // ensure isCurrentlyProfiling is true
-        // if stop is called
-            // ensure isCurrentlyProfiling is false
-            // ensure a timeout timer is not set
-            // ensure a background notification observer is not set
-    
-    // given a profiler initialized with continuous mode
-        // ensure no timeout timer is set
-        // ensure a background notification observer is set
-        // ensure isCurrentlyProfiling is true
-        // if stop is called
-            // ensure isCurrentlyProfiling is false
-            // ensure a background notification observer is not set
-    
 }
 
-private extension SentryProfilerSwiftTests {
+private extension SentryLegacyProfilerTests {
     func getLatestProfileData() throws -> Data {
         let envelope = try XCTUnwrap(self.fixture.client?.captureEventWithScopeInvocations.last)
 
@@ -588,7 +342,7 @@ private extension SentryProfilerSwiftTests {
         return try XCTUnwrap(envelope.event as? Transaction)
     }
 
-    func addMockSamples(threadMetadata: ThreadMetadata = ThreadMetadata(id: 1, priority: 2, name: "main"), addresses: [NSNumber] = [0x3, 0x4, 0x5]) throws {
+    func addMockSamples(threadMetadata: SentryProfileTestFixture.ThreadMetadata = SentryProfileTestFixture.ThreadMetadata(id: 1, priority: 2, name: "main"), addresses: [NSNumber] = [0x3, 0x4, 0x5]) throws {
         let state = try XCTUnwrap(SentryLegacyProfiler.getCurrentProfiler()).state
         fixture.currentDateProvider.advanceBy(nanoseconds: 1)
         SentryProfilerMocksSwiftCompatible.appendMockBacktrace(to: state, threadID: threadMetadata.id, threadPriority: threadMetadata.priority, threadName: threadMetadata.name, addresses: addresses)
@@ -605,7 +359,7 @@ private extension SentryProfilerSwiftTests {
 #endif // os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
     }
 
-    func performTest(transactionEnvironment: String = kSentryDefaultEnvironment, shouldTimeOut: Bool = false, uikitParameters: UIKitParameters? = nil) throws {
+    func performLegacyProfilingTest(transactionEnvironment: String = kSentryDefaultEnvironment, shouldTimeOut: Bool = false, uikitParameters: UIKitParameters? = nil) throws {
         var testingAppLaunchSpans = false
         
 #if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
@@ -634,7 +388,7 @@ private extension SentryProfilerSwiftTests {
 
         waitForExpectations(timeout: 1)
 
-        try self.assertValidProfileData(transactionEnvironment: transactionEnvironment, shouldTimeout: shouldTimeOut, appStartProfile: testingAppLaunchSpans)
+        try self.assertValidLegacyProfileData(transactionEnvironment: transactionEnvironment, shouldTimeout: shouldTimeOut, appStartProfile: testingAppLaunchSpans)
     }
 
     func assertMetricsPayload(expectedUsageReadings: Int, oneLessEnergyReading: Bool = true) throws {
@@ -720,18 +474,12 @@ private extension SentryProfilerSwiftTests {
         XCTAssertGreaterThanOrEqual(timestampNumericValue, 0)
         XCTAssertLessThanOrEqual(timestampNumericValue, transactionDuration)
     }
-
-    struct ThreadMetadata {
-        var id: UInt64
-        var priority: Int32
-        var name: String
-    }
     
     enum SentryProfilerSwiftTestError: Error {
         case notEnoughAppStartSpans
     }
 
-    func assertValidProfileData(transactionEnvironment: String = kSentryDefaultEnvironment, shouldTimeout: Bool = false, expectedAddresses: [NSNumber]? = nil, expectedThreadMetadata: [ThreadMetadata]? = nil, appStartProfile: Bool = false) throws {
+    func assertValidLegacyProfileData(transactionEnvironment: String = kSentryDefaultEnvironment, shouldTimeout: Bool = false, expectedAddresses: [NSNumber]? = nil, expectedThreadMetadata: [SentryProfileTestFixture.ThreadMetadata]? = nil, appStartProfile: Bool = false) throws {
         let data = try getLatestProfileData()
         let profile = try XCTUnwrap(try JSONSerialization.jsonObject(with: data) as? [String: Any])
